🚀 Move getDefaultUserImgId to Model as method
