+++ server/src/controllers/authController.js
@@ -8,30 +8,6 @@ const {
   attachAccessCookie,
   attachRefreshCookie,
 } = require("../utils/attachCookieTokens");
-const { getCache, setCache } = require("../services/cacheService");
-const File = require("../models/fileModel");
-
-const getDefaultUserImgId = async () => {
-  let cachedImgId = getCache("defaultUserImgId");
-  if (cachedImgId) return cachedImgId;
-
-  let defaultFile = await File.findOne({ fileId: "user" });
-
-  if (!defaultFile) {
-    defaultFile = await File.create({
-      fileId: "user",
-      name: "defaultUser.jpeg",
-      size: 0,
-      filePath: "spotify/users/defaultUser.jpeg",
-      url: "https://ik.imagekit.io/8cs4gpobr/spotify/users/defaultUser.jpeg",
-      isDefault: true,
-    });
-
-    setCache("defaultUserImgId", defaultFile.id);
-  }
-
-  return defaultFile.id;
-};
 
 exports.signUp = async (req, res, next) => {
   try {
@@ -60,7 +36,7 @@ exports.signUp = async (req, res, next) => {
     }
 
     // Set default img
-    userData.img = await getDefaultUserImgId();
+    userData.img = await User.getDefaultUserImgId();
 
     const newUser = await User.create(userData);
